# apps/backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/lessons/{id}": {
      "get": {
        "operationId": "LessonsController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      },
      "patch": {
        "operationId": "LessonsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      },
      "delete": {
        "operationId": "LessonsController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Lessons"
        ]
      }
    },
...
    "/materials/lesson/{lessonId}": {
      "get": {
        "operationId": "MaterialsController_findByLesson",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Материалы урока"
          },
          "404": {
            "description": "Урок не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить материалы урока",
        "tags": [
          "Materials"
        ]
      }
    },
...
    "/lesson-results": {
      "post": {
        "operationId": "LessonResultsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Результат урока успешно создан"
          },
          "400": {
            "description": "Некорректные данные"
          },
          "409": {
            "description": "Результат для этого студента и урока уже существует"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Выставить оценку или отметить посещаемость",
        "tags": [
          "Electronic Journal"
        ]
      },
      "get": {
        "operationId": "LessonResultsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Список всех результатов уроков"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все результаты уроков",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
    "/lesson-results/lesson/{lessonId}/journal": {
      "get": {
        "operationId": "LessonResultsController_getJournalByLesson",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "description": "ID урока",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Журнал урока со всеми студентами"
          },
          "404": {
            "description": "Урок не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить журнал по уроку (все студенты и их оценки)",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
    "/lesson-results/student/{studentId}/subject/{studyPlanId}/grades": {
      "get": {
        "operationId": "LessonResultsController_getStudentGradesBySubject",
        "parameters": [
          {
            "name": "studyPlanId",
            "required": true,
            "in": "path",
            "description": "ID учебного плана (предмета)",
            "schema": {}
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "description": "ID студента",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Оценки студента по предмету со статистикой"
          },
          "404": {
            "description": "Студент или предмет не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить все оценки студента по предмету",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
    "/lesson-results/group/{groupId}/journal": {
      "get": {
        "operationId": "LessonResultsController_getGroupJournalByPeriod",
        "parameters": [
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "description": "Дата окончания периода (YYYY-MM-DD)",
            "schema": {}
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "description": "Дата начала периода (YYYY-MM-DD)",
            "schema": {}
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "description": "ID группы",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Журнал группы за указанный период"
          },
          "404": {
            "description": "Группа не найдена"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить журнал группы за период",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
    "/lesson-results/lesson/{lessonId}/bulk-attendance": {
      "post": {
        "operationId": "LessonResultsController_bulkMarkAttendance",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "description": "ID урока",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Данные о посещаемости студентов",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "attendanceData": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "studentId": {
                          "type": "number",
                          "description": "ID студента"
                        },
                        "attendance": {
                          "type": "boolean",
                          "description": "Присутствовал ли студент"
                        },
                        "absentReason": {
                          "type": "string",
                          "enum": [
                            "SICK",
                            "FAMILY",
                            "OTHER"
                          ],
                          "description": "Причина отсутствия"
                        },
                        "absentComment": {
                          "type": "string",
                          "description": "Комментарий к отсутствию"
                        }
                      },
                      "required": [
                        "studentId",
                        "attendance"
                      ]
                    }
                  }
                },
                "required": [
                  "attendanceData"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Посещаемость успешно отмечена"
          },
          "404": {
            "description": "Урок не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Массово отметить посещаемость урока",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
    "/lesson-results/attendance/statistics": {
      "get": {
        "operationId": "LessonResultsController_getAttendanceStatistics",
        "parameters": [
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Дата окончания периода (YYYY-MM-DD, опционально)",
            "schema": {}
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Дата начала периода (YYYY-MM-DD, опционально)",
            "schema": {}
          },
          {
            "name": "studyPlanId",
            "required": false,
            "in": "query",
            "description": "ID учебного плана (опционально)",
            "schema": {}
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "description": "ID группы (опционально)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Статистика посещаемости"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить статистику посещаемости",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
    "/lesson-results/{id}": {
      "get": {
        "operationId": "LessonResultsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID результата урока",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Подробная информация о результате урока"
          },
          "404": {
            "description": "Результат урока не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Получить результат урока по ID",
        "tags": [
          "Electronic Journal"
        ]
      },
      "patch": {
        "operationId": "LessonResultsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID результата урока",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Результат урока успешно обновлен"
          },
          "404": {
            "description": "Результат урока не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Обновить результат урока (изменить оценку или посещаемость)",
        "tags": [
          "Electronic Journal"
        ]
      },
      "delete": {
        "operationId": "LessonResultsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID результата урока",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Результат урока успешно удален"
          },
          "404": {
            "description": "Результат урока не найден"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить результат урока",
        "tags": [
          "Electronic Journal"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant ABAI API",
...
    "version": "1.0.0",
    "contact": {
      "name": "ABAI Support Team",
      "url": "https://abai.edu.kz",
      "email": "support@abai.edu.kz"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.abai.edu.kz",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Введите JWT токен",
        "in": "header"
      }
    },
...
  }
}